ocr/requirements.txt
streamlit
requests
Pillow
opencv-python
numpy
python-decouple
python-dotenv
filetype